Refactor logging utilities to be shared by vipin and psimulate
